feat: order expiration

+++ src/contracts/base/Plug.Factory.sol
@@ -20,17 +20,15 @@ contract PlugFactory is PlugFactoryInterface {
     /**
      * See { PlugFactoryInterface.deploy }
      */
-    function deploy(
-        bytes calldata $salt
-    )
+    function deploy(bytes calldata $salt)
         public
         payable
         virtual
         returns (bool $alreadyDeployed, address $socketAddress)
     {
         /// @dev Recover the packed implementation and admin from the salt.
-        (uint16 nonce, address admin, address oneClicker, address implementation) =
-            abi.decode($salt, (uint16, address, address, address));
+        (uint96 nonce, address admin, address oneClicker, address implementation) =
+            abi.decode($salt, (uint96, address, address, address));
 
         /// @dev Ensure the implementation is valid.
         if (implementation == address(0) || admin == address(0)) {
@@ -38,7 +36,7 @@ contract PlugFactory is PlugFactoryInterface {
         }
 
         /// @dev Create the deployment salt using the nonce and the admin.
-        bytes32 salt = bytes32(abi.encodePacked(uint16(nonce), bytes20(admin)));
+        bytes32 salt = bytes32(abi.encodePacked(uint96(nonce), bytes20(admin)));
 
         /// @dev Deploy the new vault using a Beacon Proxy pattern.
         ($alreadyDeployed, $socketAddress) =
@@ -73,9 +71,7 @@ contract PlugFactory is PlugFactoryInterface {
     /**
      * See { PlugFactoryInterface.initCodeHash }
      */
-    function initCodeHash(
-        address $implementation
-    )
+    function initCodeHash(address $implementation)
         public
         view
         virtual
