+++ src/contracts/base/Plug.Factory.t.sol
@@ -15,23 +15,23 @@ contract PlugFactoryTest is Test {
 
     function test_salt() public {
         bytes memory salt =
-            abi.encode(uint16(1738), signer, oneClicker, address(socketImplementation));
+            abi.encode(uint96(1738), signer, oneClicker, address(socketImplementation));
 
         /// @dev Decode the details used to deploy the Socket and guard the signature.
         (
-            uint16 nonce,
+            uint96 nonce,
             address adminAddress,
             address oneClickerAddress,
             address implementationAddress
-        ) = abi.decode(salt, (uint16, address, address, address));
+        ) = abi.decode(salt, (uint96, address, address, address));
 
-        assertEq(nonce, uint16(1738));
+        assertEq(nonce, uint96(1738));
         assertEq(adminAddress, signer);
         assertEq(oneClickerAddress, oneClicker);
         assertEq(implementationAddress, address(socketImplementation));
 
         /// @dev Decode the single nonce used to guard the signature.
-        (uint16 singleNonce) = abi.decode(salt, (uint16));
+        (uint96 singleNonce) = abi.decode(salt, (uint96));
         assertEq(nonce, singleNonce);
     }
 
@@ -39,7 +39,7 @@ contract PlugFactoryTest is Test {
         vm.deal(address(this), 1000 ether);
         uint256 initialValue = _random() % 100 ether;
         bytes memory salt =
-            abi.encode(uint16(1738), signer, oneClicker, address(socketImplementation));
+            abi.encode(uint96(1738), signer, oneClicker, address(socketImplementation));
         (, address vault) = factory.deploy{ value: initialValue }(salt);
         assertEq(address(vault).balance, initialValue);
         (bool alreadyDeployed,) = factory.deploy{ value: initialValue }(salt);
@@ -47,14 +47,14 @@ contract PlugFactoryTest is Test {
     }
 
     function testRevert_InvalidImplementation(uint256) public {
-        bytes memory salt = abi.encode(uint16(1738), signer, oneClicker, address(0));
+        bytes memory salt = abi.encode(uint96(1738), signer, oneClicker, address(0));
         vm.expectRevert(abi.encodeWithSelector(PlugLib.SaltInvalid.selector, address(0), signer));
         factory.deploy(salt);
     }
 
     function testRevert_InvalidAdmin(uint256) public {
         bytes memory salt =
-            abi.encode(uint16(1738), address(0), oneClicker, address(socketImplementation));
+            abi.encode(uint96(1738), address(0), oneClicker, address(socketImplementation));
         vm.expectRevert(
             abi.encodeWithSelector(
                 PlugLib.SaltInvalid.selector, address(socketImplementation), address(0)
